+++ resources/release-notes.md
@@ -1,3 +1,5 @@
 # Software
 
+- Add `mag.mixin.Signal` as interface for signal-like classes
+- Add `resample` and `downsample` methods to `mag.IALiRT`
 - Rename `mag.mixin.Croppable` to `mag.mixin.Crop`